requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Django==4.2.3
django-appconf==1.0.5
django-imagekit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pilkit==2.0
Pillow==9.5.0
platformdirs==3.9.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
yarl==1.9.2
